Overload createElement from React
