+++ packages/element/src/react/core.js
@@ -1,0 +1,174 @@
+/**
+ * External dependencies
+ */
+// eslint-disable-next-line @typescript-eslint/no-restricted-imports
+import {
+	Children,
+	cloneElement,
+	Component,
+	createContext,
+	createRef,
+	forwardRef,
+	Fragment,
+	isValidElement,
+	memo,
+	StrictMode,
+	useCallback,
+	useContext,
+	useDebugValue,
+	useDeferredValue,
+	useEffect,
+	useId,
+	useMemo,
+	useImperativeHandle,
+	useInsertionEffect,
+	useLayoutEffect,
+	useReducer,
+	useRef,
+	useState,
+	useSyncExternalStore,
+	useTransition,
+	startTransition,
+	lazy,
+	Suspense,
+} from 'react';
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#children
+ */
+export { Children };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#cloneelement
+ */
+export { cloneElement };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#component
+ */
+export { Component };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#createcontext
+ */
+export { createContext };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#createref
+ */
+export { createRef };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#forwardref
+ */
+export { forwardRef };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#fragment
+ */
+export { Fragment };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#isvalidelement
+ */
+export { isValidElement };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#reactmemo
+ */
+export { memo };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#strictmode
+ */
+export { StrictMode };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usecallback
+ */
+export { useCallback };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usecontext
+ */
+export { useContext };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usedebugvalue
+ */
+export { useDebugValue };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usedeferredvalue
+ */
+export { useDeferredValue };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#useeffect
+ */
+export { useEffect };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#useid
+ */
+export { useId };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#useimperativehandle
+ */
+export { useImperativeHandle };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#useinsertioneffect
+ */
+export { useInsertionEffect };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#uselayouteffect
+ */
+export { useLayoutEffect };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usememo
+ */
+export { useMemo };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usereducer
+ */
+export { useReducer };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#useref
+ */
+export { useRef };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usestate
+ */
+export { useState };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usesyncexternalstore
+ */
+export { useSyncExternalStore };
+
+/**
+ * @see https://reactjs.org/docs/hooks-reference.html#usetransition
+ */
+export { useTransition };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#starttransition
+ */
+export { startTransition };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#reactlazy
+ */
+export { lazy };
+
+/**
+ * @see https://reactjs.org/docs/react-api.html#reactsuspense
+ */
+export { Suspense };
